% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\alias{run_api}
\alias{scaffold_api}
\title{API}
\usage{
run_api(host = "0.0.0.0", port = 3000)

scaffold_api()
}
\arguments{
\item{host, port}{Host and port passed to \code{plumber}.}
}
\description{
Return path to API file.
}
\details{
The \code{scaffold_api} function copies the files necessary to deploy the 
the api using docker. Make sure you do so from the directory that contains your 
configuration file (\code{\link{create_config}}). Once copied run
\code{docker build -t coronavirus .} to build the docker image with your config file.
Then run \code{docker run -p 80:8000 -d coronavirus} where \code{80} is the port you want
to map to (of your machine).
}
\section{Functions}{

\itemize{
 \item{\code{run_api} - Runs the API interactively.}
 \item{\code{scaffold_api} - Copies required files for docker deployment in current directory.}
}
}

\examples{
\dontrun{run_api()}

}
