% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_f7_table}
\alias{as_f7_table}
\title{Table}
\usage{
as_f7_table(df, card = FALSE)
}
\arguments{
\item{df}{Data.frame.}

\item{card}{Whether to use as card.}
}
\description{
Create shinyMobile table.
}
\keyword{internal}
