% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_cities_location}
\alias{china_cities_location}
\title{Coordinates}
\format{A data.frame with 3 columns and 408 rows.
\describe{
  \item{\code{cityName}}{ name of city}
  \item{\code{lat}, \code{lon}}{ coordinates}
}}
\usage{
china_cities_location
}
\description{
Coordiantes of cities.
}
\keyword{datasets}
