% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{chinese_provinces}
\alias{chinese_provinces}
\title{Dataframe to match for echarts4r geojson to work.}
\format{An object of class \code{data.frame} with 33 rows and 2 columns.}
\usage{
chinese_provinces
}
\description{
Dataframe to match for echarts4r geojson to work.
}
\keyword{datasets}
