% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect}
\alias{connect}
\alias{disconnect}
\title{Connect}
\usage{
connect()

disconnect(con = NULL)
}
\arguments{
\item{con}{Output of \code{connect}.}
}
\description{
Connect to database.
}
\keyword{internal}
