% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy}
\alias{copy}
\title{Copy URL}
\usage{
copy(embed_url, data, params)
}
\arguments{
\item{embed_url}{Base embed url.}

\item{data}{Data to use.}

\item{params}{Parameters.}
}
\description{
Copy URL
}
\keyword{internal}
