% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_iran}
\alias{covid_iran}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Iran Outbreak Dataset}
\format{
A data frame with 4 variables.
\describe{
  \item{date}{Date in YYYY-MM-DD format.}
  \item{region}{Name of region.}
  \item{province}{Name of province.}
  \item{total}{Number of confirmed cases.}
  }
}
\source{
Wikipedia contributors. 2020 coronavirus pandemic in Iran.
In Wikipedia, The Free Encyclopedia. Retrieved from
\href{https://en.wikipedia.org/w/index.php?title=2020_coronavirus_pandemic_in_Iran&oldid=951436728}{here}.
}
\usage{
covid_iran
}
\description{
Daily summary of the Coronavirus (COVID-19) confirmed cases in Iran by region
and province.
}
\details{
The dataset contains the cases in Iran by region and province.
}
\examples{
data(covid_iran)

require(dplyr)

# Get summary of total cases by province

covid_iran \%>\%
  group_by(province) \%>\%
  summarise(total = sum(cases))

}
\keyword{COVID19}
\keyword{Iran}
\keyword{coronavirus}
\keyword{datasets}
