% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_us_states}
\alias{covid_us_states}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) US Outbreak Dataset}
\format{
A data.frame object
}
\source{
The \href{https://covidtracking.com/}{COVID-19 Tracking Project website}
}
\usage{
covid_us_states
}
\description{
daily summary of the Coronavirus (COVID-19) confirmed cases in the United States by state
}
\details{
The dataset contains the daily summary of the Coronavirus confirmed cases in the US by state
}
\examples{

data(covid_us_states)

library(dplyr)

# Get summary by sub region of totalcases

covid_us_states
}
\keyword{COVID19}
\keyword{US}
\keyword{coronavirus}
\keyword{datasets}
