% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl_coronavirus}
\alias{crawl_coronavirus}
\title{Crawl Data}
\usage{
crawl_coronavirus()
}
\description{
Crawl John Hopkin's data and store in database. 
Every time the function is called the entirety of 
the data is pulled and the content of the database 
is overwritten.
}
