% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geoloc_dxy}
\alias{geoloc_dxy}
\title{Geolocate DXY Data}
\usage{
geoloc_dxy(df)
}
\description{
Geolocate DingXiangYuan data.
}
\details{
This is used internally to create database of geolocated Chinese cities.
}
\keyword{internal}
