% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_coronvirus.R
\name{get_info_coronavirus}
\alias{get_info_coronavirus}
\title{Get information about the datasets provided by the coronavirus package}
\usage{
get_info_coronavirus()
}
\value{
a tibble of information about the datasets in this package
}
\description{
Returns information about the datasets in this package for covid19R harvesting
}
\examples{
\dontrun{

# get the dataset info from this package
get_info_coronavirus()
}

}
