% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeds_server.R
\name{get_query}
\alias{get_query}
\title{Get Query}
\usage{
get_query(query, param)
}
\arguments{
\item{query}{URL query.}

\item{param}{Parameter to extract.}
}
\description{
Get Query
}
\keyword{internal}
