% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_city_map.R
\name{interface}
\alias{interface}
\alias{input_to_case}
\alias{input_to_palette}
\title{Input to Case Conversion}
\usage{
input_to_case(x)

input_to_palette(x)
}
\arguments{
\item{x}{input value.}
}
\description{
Converts case name input to dxy column name
}
\keyword{internal}
