% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_china_others.R
\name{mod_china_others_ui}
\alias{mod_china_others_ui}
\alias{mod_china_others_server}
\title{mod_china_others_ui and mod_china_others_server}
\usage{
mod_china_others_ui(id)

mod_china_others_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
