% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_china_trend.R
\name{mod_china_trend_ui}
\alias{mod_china_trend_ui}
\alias{mod_china_trend_server}
\title{mod_chiny_trend_ui and mod_chiny_trend_server}
\usage{
mod_china_trend_ui(id, label)

mod_china_trend_server(
  input,
  output,
  session,
  df,
  column = "confirm",
  connect = FALSE
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
