% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_count_weixin.R
\name{mod_count_weixin_ui}
\alias{mod_count_weixin_ui}
\alias{mod_count_weixin_server}
\title{mod_count_weixin_ui and mod_count_weixin_server}
\usage{
mod_count_weixin_ui(id, label = NULL, source = "Weixin", color = "#fff")

mod_count_weixin_server(input, output, session, df, column)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
