% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dash.R
\name{mod_dash_ui}
\alias{mod_dash_ui}
\alias{mod_dash_server}
\title{mod_dash_ui and mod_dash_server}
\usage{
mod_dash_ui(id)

mod_dash_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
