% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dxy_table.R
\name{mod_dxy_table_ui}
\alias{mod_dxy_table_ui}
\alias{mod_dxy_table_server}
\title{mod_dxy_table_ui and mod_dxy_table_server}
\usage{
mod_dxy_table_ui(id)

mod_dxy_table_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
