% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geo_lines.R
\name{mod_geo_lines_ui}
\alias{mod_geo_lines_ui}
\alias{mod_geo_lines_server}
\title{mod_geo_lines_ui and mod_geo_lines_server}
\usage{
mod_geo_lines_ui(id)

mod_geo_lines_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
