% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_jhu_death_rate.R
\name{mod_jhu_death_rate_ui}
\alias{mod_jhu_death_rate_ui}
\alias{mod_jhu_death_rate_server}
\title{mod_jhu_death_rate_ui and mod_jhu_death_rate_server}
\usage{
mod_jhu_death_rate_ui(id)

mod_jhu_death_rate_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
