% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_new_cases.R
\name{mod_new_cases_ui}
\alias{mod_new_cases_ui}
\alias{mod_new_cases_server}
\title{mod_new_cases_ui and mod_new_cases_server}
\usage{
mod_new_cases_ui(id)

mod_new_cases_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
