% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_news.R
\name{mod_news_ui}
\alias{mod_news_ui}
\alias{mod_news_server}
\title{mod_news_ui and mod_news_server}
\usage{
mod_news_ui(id)

mod_news_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
