% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_china_ui}
\alias{mod_china_ui}
\alias{mod_china_server}
\title{mod_table_ui and mod_table_server}
\usage{
mod_china_ui(id, label)

mod_china_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
