% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_time_provinces.R
\name{mod_time_provinces_ui}
\alias{mod_time_provinces_ui}
\alias{mod_time_provinces_server}
\title{mod_time_provinces_ui and mod_time_provinces_server}
\usage{
mod_time_provinces_ui(id)

mod_time_provinces_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
