% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trend.R
\name{mod_trend_ui}
\alias{mod_trend_ui}
\alias{mod_trend_server}
\title{mod_trend_ui and mod_trend_server}
\usage{
mod_trend_ui(id)

mod_trend_server(input, output, session, df = df, type_filter = "confirmed")
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
