% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_world.R
\name{mod_world_ui}
\alias{mod_world_ui}
\alias{mod_world_server}
\title{mod_world_ui and mod_world_server}
\usage{
mod_world_ui(id)

mod_world_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
