% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_up}
\alias{round_up}
\title{Round up}
\usage{
round_up(x, nice = 1:10)
}
\arguments{
\item{x}{Number.}

\item{nice}{Vector of ending.}
}
\description{
Nicely round up numbers for pieces in maps.
}
\keyword{internal}
