% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_coronavirus.R
\name{update_coronavirus}
\alias{update_coronavirus}
\title{update_coronavirus}
\source{
Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}
}
\usage{
update_coronavirus()
}
\value{
An updated tbl object for the coronavirus dataset.
}
\description{
Update the data using the daily summary of the Coronavirus (COVID-19) cases by state/province.
}
\examples{
\dontrun{
coronavirus <- update_coronavirus()
}
}
