% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_coronavirus_cds.R
\docType{data}
\name{update_coronavirus_cds_raw}
\alias{update_coronavirus_cds_raw}
\title{Update the coronavirus CDS Datasets}
\format{
An object of class \code{function} of length 1.
}
\source{
\href{https://coronadatascraper.com/#home}{website}
}
\usage{
update_coronavirus_cds_raw
}
\arguments{
\item{remake_table}{Defaults to FALSE. If TRUE, remakes the geocode fix table}

\item{geofix}{fix the bad lat/longs?}

\item{fixed_geotable}{if you have an alternate data.frame to fix the geospatial errors, supply it.}

\item{returnclass}{Return an sf or a data frame object}

\item{googlemaps_api_key}{If you are fixing the geotable, you need a googlemaps api key. See \link[ggmaps](register_google) for more.
Defaults to NULL.}
}
\value{
A data.frame or sf object
}
\description{
Update the coronavirus data scrapper datasets
}
\details{
As the CRAN version is being updated every one-two months, the dev version of the package is being updated on a daily bases.
This function enables to refresh the package directly from the source with the most up-to-date data
}
\keyword{datasets}
