% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_coronavirus_raw.R
\name{update_coronavirus_raw}
\alias{update_coronavirus_raw}
\title{update_coronavirus_raw}
\source{
Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}

The \link[rnaturalearth]{rnaturalearth}
}
\usage{
update_coronavirus_raw(debug = FALSE, returnclass = c("data.frame", "sf"))
}
\arguments{
\item{debug}{Should we print out the head of datasets as they are downloaded or other debug
information? Defaults to `FALSE`.}

\item{returnclass}{Return}
}
\value{
An updated tbl or sf object for the coronavirus dataset.
}
\description{
Update the data using the daily summary of the Coronavirus (COVID-19) cases by
state/province. Note, this function pulls from the raw archive. The data has not been
QC-ed yet on our end, and as such might have some innaaccuracies in it, such as duplicated
records, negative values, or other issues that we have noticed in the past. Please file issues
at the data source - https://github.com/CSSEGISandData/COVID-19/ - and you can give us a heads
up at https://github.com/RamiKrispin/coronavirus/issues as we will be pulling it soon.

Spatial information (borders, etc.) are from \link[rnaturalearth]{rnaturalearth}. `sf` objects
can be easily merged with other data from \link[coronavirus]{coronavirus} using \link[sf]{st_join}.
}
\examples{
\dontrun{
coronavirus <- update_coronavirus_raw()
}
}
