% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{update_dataset}
\alias{update_dataset}
\title{Update the coronavirus Dataset}
\source{
coronavirus - Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}
}
\usage{
update_dataset(silence = FALSE)
}
\arguments{
\item{silence}{A boolean, if set to TRUE, will automatically install updates without prompt question, by default set to FALSE}
}
\value{
A data.frame object
}
\description{
Update the package datasets on the global environment with the most recent data on the Dev version
}
\details{
As the CRAN version is being updated every one-two months, the dev version of the package is being updated on a daily bases.
This function enables to refresh the package dataset to the most up-to-date data. Changes will be available on the global environment
}
