% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{update_datasets}
\alias{update_datasets}
\title{Update the Package Datasets}
\source{
coronavirus - Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}

covid_italy - Wikipedia article "2020 coronavirus outbreak in Italy" \href{https://en.wikipedia.org/wiki/2020_coronavirus_outbreak_in_Italy}{website}

covid_south_korea - Wikipedia article "2020 coronavirus outbreak in South Korea" \href{https://en.wikipedia.org/wiki/2020_coronavirus_outbreak_in_South_Korea}{website}

covid_iran - Wikipedia article "2020 coronavirus pandemic in Iran"  \href{https://en.wikipedia.org/wiki/2020_coronavirus_pandemic_in_Iran}{website}
}
\usage{
update_datasets()
}
\value{
A data.frame object
}
\description{
Update the package datasets on the global environment with the most recent data on the Dev version
}
\details{
As the CRAN version is being updated every one-two months, the dev version of the package is being updated on a daily bases.
This function enables to refresh the package dataset to the most up-to-date data. Changes will be available on the global environment
}
\examples{
\dontrun{
update_datasets()
}
}
