% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_city_map.R
\name{url_to_geojson}
\alias{url_to_geojson}
\title{County Geojson}
\usage{
url_to_geojson(province)
}
\arguments{
\item{province}{Province pinyin.}
}
\description{
Get url to a county's geojson.
}
\keyword{internal}
