% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_population}
\alias{world_population}
\title{World Population by Country Dataset}
\format{
A data frame with 4 variables.
\describe{
  \item{country_name}{Country Name}
  \item{country_code}{Country abbreviations code}
  \item{year}{The year of the observation}
  \item{population}{Population}
  }
}
\source{
The World Bank Data \href{https://data.worldbank.org/indicator/SP.POP.TOTL?most_recent_year_desc=false}{website}.
}
\usage{
world_population
}
\description{
World population by country between 1960 to 2020
}
\details{
The dataset provides worldwide popluation values by country between 1960 and 2020
}
\examples{

library(dplyr)
data(world_population)

head(world_population)

us_pop <- world_population \%>\%
  filter(country_code == "USA")
plot(us_pop$year, us_pop$population,
     type = "l", main = "US Population",
     xlab = "Source: World Bank",
     ylab = "Population",
     col = "blue")
}
\keyword{country}
\keyword{datasets}
\keyword{population}
