% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinese_provinces}
\alias{chinese_provinces}
\title{China Provinces}
\format{A data.frame with 2 columns and 33 rows.
\describe{
  \item{\code{state}}{ name of province}
  \item{\code{chinese}}{ translation}
}}
\usage{
chinese_provinces
}
\description{
Chinese provinces and their translations.
}
\keyword{datasets}
