% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{Configuration}
\usage{
create_config()
}
\description{
Creates a configuration file.
}
\section{newsapi}{

The latest version requires a newsapi token from \url{https://newsapi.org/}.
}

