% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_iran}
\alias{covid_iran}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Iran Outbreak Dataset}
\format{
A data.frame object
}
\source{
Wikipedia article "2020 coronavirus pandemic in Iran"  \href{https://en.wikipedia.org/wiki/2020_coronavirus_pandemic_in_Iran}{website}
}
\usage{
covid_iran
}
\description{
daily summary of the Coronavirus (COVID-19) confirmed cases in Iran by region and province
}
\details{
The dataset contains the cases in Iran by region and province
}
\examples{

data(covid_iran)

require(dplyr)

# Get summary of total cases by province

covid_iran \%>\%
group_by(province) \%>\%
summarise(total = sum(cases))

}
\keyword{COVID19}
\keyword{Iran}
\keyword{coronavirus}
\keyword{datasets}
