% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_italy}
\alias{covid_italy}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Italy Outbreak Dataset}
\format{A data.frame object}
\source{
Presidenza del Consiglio dei Ministri - Dipartimento della Protezione Civile \href{http://www.protezionecivile.it/}{website}
}
\usage{
covid_italy
}
\description{
daily summary of the Coronavirus (COVID-19) confirmed cases in Italy by province
}
\details{
The dataset contains the daily summary of the Coronavirus confirmed cases in Italy by province
}
\examples{

data(covid_italy)

require(dplyr)

# Get summary by sub region of totalcases

covid_italy \%>\%
group_by(sub_region) \%>\%
summarise(total_cases = sum(total))

}
\keyword{COVID19}
\keyword{Italy}
\keyword{coronavirus}
\keyword{datasets}
