% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_south_korea}
\alias{covid_south_korea}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) South Korea Outbreak Dataset}
\format{
A data.frame object
}
\source{
Wikipedia article "2020 coronavirus outbreak in South Korea" \href{https://en.wikipedia.org/wiki/2020_coronavirus_outbreak_in_South_Korea}{website}
}
\usage{
covid_south_korea
}
\description{
daily summary of the Coronavirus (COVID-19) confirmed cases in South Korea by province and city
}
\details{
The dataset contains the daily summary of the Coronavirus confirmed cases in South Korea by province and city
}
\examples{

data(covid_south_korea)

require(dplyr)

# Get summary of total cases by city

covid_south_korea \%>\%
group_by(city) \%>\%
summarise(total_cases = sum(total))

}
\keyword{COVID19}
\keyword{Korea}
\keyword{South}
\keyword{coronavirus}
\keyword{datasets}
