% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refresh.R
\name{update_datasets}
\alias{update_datasets}
\title{Update the Package Datasets}
\source{
Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}
}
\usage{
update_datasets()
}
\value{
A data.frame object
}
\description{
Update the package datasets from the Dev version
}
\details{
As the CRAN version is being updated every one-two months, the dev version of the package is being updated on a daily bases.
This function enables to refresh the package dataset to the most up-to-date data
}
\examples{
\dontrun{
update_datasets()
}
}
